+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/DoctorClinicService.java
@@ -16,6 +16,6 @@ public interface DoctorClinicService {
 
     DoctorClinic getDoctorClinicFromDoctorAndClinic(Doctor doctor, Clinic clinic);
 
-    List<DoctorClinic> getDoctorBy(Location location, Specialty specialty, long clinic);
+    List<DoctorClinic> getDoctorBy(Location location, Specialty specialty, int clinic);
 
 }
